compiling on linux

+++ src/scenes/space-invaders/player.h
@@ -8,5 +8,5 @@ namespace mg {
 
 struct Player;
 
-void transformPlayer(Player *player, uint32_t screenWidth, uint32_t spriteSize, const mg::FrameData &frameData, float dt);
+void transformPlayer(Player *player, uint32_t screenWidth, float spriteSize, const mg::FrameData &frameData, float dt);
 bool shouldFirePlayerBullet(Player *player, const mg::FrameData &frameData, float dt);
